+++ src/Bot.ts
@@ -23,6 +23,8 @@ export class Bot extends Client {
             prefix: '!'
         };
 
+        this.config = Object.assign({}, defaultConfig, config);
+
         this.addCommand({
             name: 'help',
             description: 'Print this help message',
@@ -30,7 +32,6 @@ export class Bot extends Client {
             requiredArgCount: 0
         });
 
-        this.config = Object.assign({}, defaultConfig, config);
         this.on('message', this.handleMessage.bind(this));
     }
 
@@ -40,7 +41,7 @@ export class Bot extends Client {
         return this.commands.some(command => {
             const parser = new Parser(message.content);
 
-            if (parser.getCommandName() === `${this.config.prefix}${command.name}`) {
+            if (parser.getCommandName() === command.name) {
                 if (!command.requiredArgCount || parser.getArgsCount() >= command.requiredArgCount) {
                     command.action(message, parser.getArgs());
                 } else {
@@ -56,7 +57,14 @@ export class Bot extends Client {
         const spacePaddingLength = 9;
         let helpMessage = '```';
         helpMessage += 'Available commands:\n\n';
-        helpMessage += this.commands.map(command => `${this.config.prefix}${command.name}${' '.repeat(spacePaddingLength - command.name.length)}${command.description}`).join('\n');
+        helpMessage += this.commands.map(command => {
+            let str = '';
+            str += `${command.name}`;
+            str += `${' '.repeat(spacePaddingLength - command.name.length)}`;
+            str += `${command.description ?? ''}`;
+            str += command.usage ? `\n${' '.repeat(spacePaddingLength)}Usage: ${command.usage}` : '';
+            return str;
+        }).join('\n');
         helpMessage += '```';
 
         Bot.replyToMessage(message, helpMessage);
@@ -67,6 +75,10 @@ export class Bot extends Client {
     }
 
     addCommand(command: Command): Bot {
+        if(!command.name.startsWith(this.config.prefix)){
+            command.name = this.config.prefix + command.name
+        }
+
         this.commands.push(command);
         return this;
     }