+++ src/Command.ts
@@ -2,7 +2,8 @@ import {Message} from "discord.js";
 
 export interface Command {
     name: string;
-    description: string;
+    description?: string;
+    usage?: string;
     action: (message: Message, args: string[]) => boolean | void;
     requiredArgCount?: number;
 }