+++ CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Next
+#### Added
+- Usage in command object for help message
+
+#### Changed
+- Command name can start with prefix or not. For example, with prefix `!`, command name can be `ping` or `!ping`
+
 ### 1.1.3
-### Added
+#### Added
 - Added [CODE_OF_CONDUCT.md](CODE_OF_CONDUCT.md)
 - Added [CONTRIBUTING.md](CONTRIBUTING.md)
 - Added [CHANGELOG.md](CHANGELOG.md)